backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
ultralytics
opencv-python-headless
pillow
google-generativeai
python-dotenv